resolved issues #1 #3 from 5
